// src/api/clients/kyc/constants.ts
export enum RejectionReasons {
  missingDocumentPhoto = 'MissingDocumentPhoto',
  documentNotAccepted = 'DocumentNotAccepted',
  dataDoesNotMatch = 'DataDoesNotMatch',
  sanctioned = 'Sanctioned',
  missingSelfiePhoto = 'MissingSelfiePhoto',
  facesDoesNotMatch = 'FacesDoesNotMatch',
  poorPhotoQuality = 'PoorPhotoQuality',
  poorPhotoLighting = 'PoorPhotoLighting',
  blurredPhoto = 'BlurredPhoto',
  badMediaFormat = 'BadMediaFormat',
  miscellaneous = 'Miscellaneous',
  possibleFraudAttempt = 'PossibleFraudAttempt',
  unrelatedPhoto = 'UnrelatedPhoto',
  moreThanOnePerson = 'MoreThanOnePerson',
  prohibitedCountryOrState = 'ProhibitedCountryOrState',
  documentIsExpired = 'DocumentIsExpired',
  documentWithNonLatinCharacters = 'DocumentWithNonLatinCharacters',
  partOfDocumentIsCovered = 'PartOfDocumentIsCovered',
  nameMismatch = 'NameMismatch',
  expirationDateMismatch = 'ExpirationDateMismatch',
  incorrectInformationProvided = 'IncorrectInformationProvided',
  partOfFaceIsCovered = 'PartOfFaceIsCovered',
  underagePerson = 'UnderagePerson',
  prohibitedNationality = 'ProhibitedNationality',
  docAndBioAgeDoesntMatch = 'DocAndBioAgeDoesntMatch',
  duplicatedInfo = 'DuplicatedInfo',
}

export const CriticalReasons: RejectionReasons[] = [
  RejectionReasons.sanctioned,
  RejectionReasons.miscellaneous,
  RejectionReasons.possibleFraudAttempt,
  RejectionReasons.prohibitedCountryOrState,
  RejectionReasons.underagePerson,
  RejectionReasons.prohibitedNationality,
  RejectionReasons.docAndBioAgeDoesntMatch,
  RejectionReasons.duplicatedInfo,
];